+++ packages/e2e/cypress/plugins/index.ts
@@ -26,5 +26,4 @@ const Plugin: Cypress.PluginConfig = (on, config) => {
   return config;
 };
 
-// eslint-disable-next-line import/no-default-export
 export default Plugin;