+++ packages/scripts/src/commands/release.ts
@@ -1,7 +1,7 @@
 import { promises as fs } from 'fs';
 import { basename, dirname, join, parse } from 'path';
 
-import { getWorkspaces, logger, opendirSafe, readYaml } from '@appsemble/node-utils';
+import { getWorkspaces, logger, opendirSafe, readData, writeData } from '@appsemble/node-utils';
 import { AppsembleMessages } from '@appsemble/types';
 import { formatISO } from 'date-fns';
 import { ensureFile, readJson, remove, writeJson } from 'fs-extra';
@@ -15,7 +15,6 @@ import * as semver from 'semver';
 import { PackageJson } from 'type-fest';
 import { Argv } from 'yargs';
 
-import { outputYaml } from '../lib/fs';
 import {
   createHeading,
   createLink,
@@ -83,10 +82,10 @@ async function updatePkg(dir: string, version: string): Promise<void> {
  * @param version - The software version to set
  */
 async function updatePublicCodeYml(version: string): Promise<void> {
-  const [publicCode] = await readYaml<any>('publiccode.yml');
+  const [publicCode] = await readData<any>('publiccode.yml');
   const i18nFiles = await fs.readdir('i18n');
   const availableLanguages = i18nFiles.map((f) => parse(f).name).sort();
-  await outputYaml(
+  await writeData(
     'publiccode.yml',
     mapValues(publicCode, (value, key) => {
       switch (key) {
@@ -103,6 +102,7 @@ async function updatePublicCodeYml(version: string): Promise<void> {
           return value;
       }
     }),
+    { sort: false },
   );
 }
 
@@ -199,7 +199,7 @@ async function updateChangelog(changesByCategory: Changes, version: string): Pro
 }
 
 async function updateHelmChart(changes: Changes, version: string): Promise<void> {
-  const [chart] = await readYaml<any>('config/charts/appsemble/Chart.yaml');
+  const [chart] = await readData<any>('config/charts/appsemble/Chart.yaml');
   const changelog = dump(
     Object.entries(changes).flatMap(([kind, entries]) =>
       entries.map((entry: ListItem) => ({
@@ -208,7 +208,7 @@ async function updateHelmChart(changes: Changes, version: string): Promise<void>
       })),
     ),
   );
-  await outputYaml(
+  await writeData(
     'config/charts/appsemble/Chart.yaml',
     mapValues(chart, (value, key) => {
       switch (key) {
@@ -221,6 +221,7 @@ async function updateHelmChart(changes: Changes, version: string): Promise<void>
           return value;
       }
     }),
+    { sort: false },
   );
 }